server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
arabic-reshaper==3.0.0
    # via xhtml2pdf
arrow==0.13.0
    # via
    #   eve-elastic
    #   superdesk-core
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   pyhanko
    #   pyhanko-certvalidator
async-timeout==4.0.3
    # via redis
authlib==0.14.3
    # via superdesk-core
babel==2.14.0
    # via flask-babel
bcrypt==3.1.7
    # via superdesk-core
billiard==3.6.4.0
    # via celery
blinker==1.7.0
    # via
    #   elastic-apm
    #   flask-mail
    #   raven
    #   sentry-sdk
    #   superdesk-core
boto3==1.34.84
    # via superdesk-core
botocore==1.34.84
    # via
    #   boto3
    #   s3transfer
cachelib==0.9.0
    # via flask-caching
cachetools==5.3.3
    # via
    #   flask-oidc-ex
    #   google-auth
celery[redis]==5.2.7
    # via superdesk-core
cerberus==1.3.5
    # via
    #   eve
    #   superdesk-core
certifi==2024.2.2
    # via
    #   elastic-apm
    #   elasticsearch
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   bcrypt
    #   cryptography
chardet==5.2.0
    # via superdesk-core
charset-normalizer==3.3.2
    # via requests
ciso8601==2.3.1
    # via eve-elastic
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
    #   newsroom-core
    #   pyhanko
    #   superdesk-core
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
croniter==0.3.37
    # via superdesk-core
cryptography==42.0.5
    # via
    #   authlib
    #   jwcrypto
    #   pyhanko
    #   pyhanko-certvalidator
cssselect2==0.7.0
    # via svglib
deepdiff==7.0.1
    # via superdesk-planning
deprecated==1.2.14
    # via limits
draftjs-exporter[lxml]==2.1.7
    # via superdesk-core
ecs-logging==2.1.0
    # via elastic-apm
elastic-apm[flask]==6.22.0
    # via superdesk-core
elasticsearch==7.13.4
    # via
    #   eve-elastic
    #   superdesk-core
eve==1.1.2
    # via superdesk-core
eve-elastic==7.4.1
    # via superdesk-core
events==0.3
    # via eve
feedparser==6.0.11
    # via superdesk-core
flask==1.1.2
    # via
    #   eve
    #   flask-babel
    #   flask-caching
    #   flask-limiter
    #   flask-mail
    #   flask-oidc-ex
    #   flask-pymongo
    #   flask-script
    #   flask-webpack
    #   flask-wtf
    #   newsroom-core
    #   raven
    #   sentry-sdk
    #   superdesk-core
flask-babel==1.0.0
    # via superdesk-core
flask-caching==2.1.0
    # via newsroom-core
flask-limiter==0.9.5.1
    # via newsroom-core
flask-mail==0.9.1
    # via superdesk-core
flask-oidc-ex==0.5.5
    # via superdesk-core
flask-pymongo==0.5.2
    # via newsroom-core
flask-script==2.0.6
    # via superdesk-core
flask-webpack==0.1.0
    # via newsroom-core
flask-wtf==0.14.3
    # via newsroom-core
future==1.0.0
    # via python-twitter
google-auth==2.6.6
    # via newsroom-core
gunicorn==20.0.4
    # via newsroom-core
hachoir==3.0a3
    # via superdesk-core
hermescache==0.10.0
    # via superdesk-core
honcho==1.1.0
    # via newsroom-core
html5lib==1.1
    # via xhtml2pdf
httplib2==0.22.0
    # via oauth2client
icalendar==4.0.9
    # via superdesk-planning
idna==3.7
    # via requests
importlib-resources==6.4.0
    # via limits
isodate==0.6.1
    # via python3-saml
itsdangerous==1.1.0
    # via
    #   flask
    #   flask-oidc-ex
    #   flask-wtf
    #   superdesk-core
jinja2==2.11.3
    # via
    #   flask
    #   flask-babel
    #   superdesk-core
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jwcrypto==1.5.6
    # via
    #   flask-oidc-ex
    #   python-jwt
kombu==5.2.4
    # via
    #   celery
    #   superdesk-core
ldap3==2.5.2
    # via superdesk-core
limits==3.10.1
    # via flask-limiter
lxml==4.6.5
    # via
    #   draftjs-exporter
    #   python3-saml
    #   superdesk-core
    #   svglib
    #   xmlsec
markupsafe==2.0.1
    # via
    #   jinja2
    #   newsroom-core
    #   superdesk-core
mongolock==1.3.4
    # via superdesk-core
natsort==8.4.0
    # via croniter
newsroom-core @ git+https://github.com/superdesk/newsroom-core.git@release/2.6
    # via -r requirements.in
oauth2client==4.1.3
    # via flask-oidc-ex
oauthlib==3.2.2
    # via requests-oauthlib
ordered-set==4.1.0
    # via deepdiff
oscrypto==1.3.0
    # via pyhanko-certvalidator
packaging==24.0
    # via limits
pillow==9.2.0
    # via
    #   reportlab
    #   superdesk-core
    #   xhtml2pdf
prompt-toolkit==3.0.43
    # via click-repl
pyasn1==0.6.0
    # via
    #   ldap3
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   oauth2client
pycparser==2.22
    # via cffi
pyhanko==0.18.0
    # via xhtml2pdf
pyhanko-certvalidator==0.22.0
    # via
    #   pyhanko
    #   xhtml2pdf
pyjwt==2.4.0
    # via superdesk-core
pymemcache==4.0.0
    # via superdesk-core
pymongo==3.11.4
    # via
    #   eve
    #   flask-pymongo
    #   mongolock
    #   superdesk-core
pyparsing==3.1.2
    # via
    #   httplib2
    #   pyrtf3
pypdf==4.2.0
    # via xhtml2pdf
pypng==0.20220715.0
    # via qrcode
pyrtf3==0.47.5
    # via newsroom-core
python-bidi==0.4.2
    # via xhtml2pdf
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   croniter
    #   icalendar
    #   superdesk-core
python-jwt==4.1.0
    # via flask-oidc-ex
python-magic==0.4.27
    # via superdesk-core
python-twitter==3.5
    # via superdesk-core
python3-saml==1.15.0
    # via newsroom-core
pytz==2024.1
    # via
    #   celery
    #   eve-elastic
    #   flask-babel
    #   icalendar
    #   pyhanko
    #   superdesk-core
    #   tzlocal
pyyaml==6.0.1
    # via
    #   pyhanko
    #   superdesk-core
qrcode==7.4.2
    # via pyhanko
raven[flask]==6.10.0
    # via superdesk-core
redis==4.5.5
    # via
    #   celery
    #   superdesk-core
regex==2020.7.14
    # via superdesk-core
reportlab==3.6.13
    # via
    #   newsroom-core
    #   superdesk-core
    #   svglib
    #   xhtml2pdf
requests==2.31.0
    # via
    #   pyhanko
    #   pyhanko-certvalidator
    #   python-twitter
    #   requests-oauthlib
    #   superdesk-core
requests-oauthlib==2.0.0
    # via python-twitter
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.10.1
    # via boto3
sentry-sdk[flask]==1.5.12
    # via newsroom-core
sgmllib3k==1.0.0
    # via feedparser
simplejson==3.19.2
    # via eve
six==1.16.0
    # via
    #   bcrypt
    #   flask-limiter
    #   flask-oidc-ex
    #   google-auth
    #   html5lib
    #   isodate
    #   oauth2client
    #   python-bidi
    #   python-dateutil
superdesk-core @ git+https://github.com/superdesk/superdesk-core.git@v2.7.0-rc4
    # via newsroom-core
superdesk-planning @ git+https://github.com/superdesk/superdesk-planning.git@v2.7.0-rc1
    # via newsroom-core
svglib==1.5.1
    # via xhtml2pdf
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.11.0
    # via
    #   jwcrypto
    #   limits
    #   pypdf
    #   qrcode
    #   superdesk-core
tzlocal==2.1
    # via
    #   pyhanko
    #   superdesk-core
unidecode==0.4.21
    # via superdesk-core
uritools==4.0.2
    # via pyhanko-certvalidator
urllib3==1.25.11
    # via
    #   botocore
    #   elastic-apm
    #   elasticsearch
    #   requests
    #   sentry-sdk
    #   superdesk-core
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
websockets==10.3
    # via superdesk-core
werkzeug==1.0.1
    # via
    #   flask
    #   superdesk-core
wrapt==1.14.1
    # via
    #   deprecated
    #   elastic-apm
wtforms==2.2.1
    # via
    #   flask-wtf
    #   newsroom-core
xhtml2pdf==0.2.11
    # via newsroom-core
xmlsec==1.3.13
    # via python3-saml

# The following packages are considered to be unsafe in a requirements file:
# setuptools
